Update migrations in tests app
